feat(sdk): remove logger
The `Logger` resource basically doesn't add value, only complexity. It is a great example of redundant future proofing, so let's just get rid of it and make the word a better place! We can always introduce it later when use cases are clearer.

Let me explain: the only way to emit a log right now in Wing is using the `log()` builtin function, which is jsified as a `console.log()` statement (both preflight and inflight). Now, we hijacked `console.log` in our runtime environments and mapped it to the standard logger of that environment.

In the cloud (e.g. AWS Lambda), the "standard logger" is basically `console.log()`, so basically the implementation of the logger for these environments is a "no-op".

In the simulator, since https://github.com/winglang/wing/pull/1879, we can listen to `console.xxx` events that come from the VM, and map them directly to the simulator trace system, so there's no need to hijack `console.log` here as well.

All in all, doctor says it's a classic (and severe) case of future proofing, where we get the exact same behavior with much less complexity and code to maintain (over 700 lines of code are not needed) ***== GOODNESS**.

## Misc

`node-fetch@3` is an ESM-only module and cannot be imported into our VM (damn!). In the meantime, I've downgraded it to 2.x and opened a bug: https://github.com/winglang/wing/issues/2059