fix: unable to resolve extern depending on cwd (#2083)
Fixes #2081

When running wing from the same directory as the wing file, relative extern specs failed to be found in wingc. This was because the WASI preopens didn't understand the given absolute path when when trying to read the same directory as ".". To be completely honest, I'm not sure why it worked from a parent directory.

I added this test to apps/wing instead of hangar, and while I was there I changed it to use vitest instead of jest since the rest of our repo uses it.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.