feat!: expose resource (and construct) apis (WIP)
Moved the `Resource` class from `core.Resource` (which is a hidden namespace) to `std.Resource` (which is for visible built-in types).

Removed much of the special-casing related to short-circuiting `Construct` (resolves #2028) and `Resource` during the jsii import and type checking.

To resolve #1595, if a `resource` is defined without a base class, implicitly associate `std.Resource` as a parent.

Moved the JSII `TypeSystem` object to be shared across JSII imports. This object is designed to be used as a mutable global cache for imported JSII types. Previously, every time a module was imported under a different alias, we reloaded the JSII typesystem, which is not needed.

As a follow up for #2076 in order to simplify the jsii importer so we can simply import all dependencies of the SDK, we have converted all the target-specific SDK dependencies to bundled dependencies instead of normal deps. This is fine because none of these types are exposed in the SDK's API, so JSII is happy. This dramatically decreases the public dependency surface area of the SDK and later on we can follow up with installing these dependencies just in time to reduce the package size (related to #2035).

Simplified wingii APIs by removing the concept of `roots` (not used) and the option to not load `deps`. This is not needed now because our SDK is only dependent on `constructs` which we want to load (because we want to allow folks to access its API).

TODO:

- [ ] Remove a million debug prints
- [ ] Allow referencing `std.Resource` as `Resource` (or `resource`)?
- [ ] Cache the `std.Resource` so we don't need to lookup every time (`type_check.rs:2101`).
- [ ] Determine if we need the root namespace in `import_submodules_to_env`.
- [ ] Make all the types under `std` accessible via their name without a namespace (??).
- [ ] Fix this:

```
error: Expected type to be "IConstruct", but got "WingResource" instead
   --> bring-cdk.w:29:16
   |
29 | log(cx.Node.of(wr).path);
   |                ^^ Expected type to be "IConstruct", but got "WingResource" instead
   ```

Example:

```js
bring cloud;
bring "constructs" as cx;
bring "aws-cdk-lib" as aws;

resource WingResource {
  init() {
    log("my id is ${this.node.id}");
  }
}

let get_path = (c: cx.Construct): str => {
  return c.node.path;
};

let get_display_name = (r: std.Resource): str? => {
  return r.display.title;
};

let q = new aws.aws_sqs.Queue();
let wr = new WingResource();
let another_resource: std.Resource = wr;

log("path of sqs.queue: ${get_path(q)}");
log("path of wing resource: ${get_path(wr)}");

let title = get_display_name(wr) ?? "no display name";
log("display name of wing resource: ${title}");

log(cx.Node.of(wr).path);
```