feat!: expose base resource and construct apis (#2095)
Moved the `Resource` class from `core.Resource` (which is a hidden namespace) to `std.Resource` (which is for visible built-in types).

Removed much of the special-casing related to short-circuiting `Construct` (resolves #2028) and `Resource` during the jsii import and type checking.

To resolve #1595, if a `resource` is defined without a base class, implicitly associate `std.Resource` as a parent.

Moved the JSII `TypeSystem` object to be shared across JSII imports. This object is designed to be used as a mutable global cache for imported JSII types. Previously, every time a module was imported under a different alias, we reloaded the JSII typesystem, which is not needed. This also means that we don't need to persist `jsii_imports` because they are now basically just a set of specs and there isn't much value in persisting them across `TypeChecker` instantiations (@MarkMcCulloh can you validate this assumption please?).

As a follow up for #2076 in order to simplify the jsii importer so we can simply import all dependencies of the SDK, we have converted all the target-specific SDK dependencies to bundled dependencies instead of normal deps. This is fine because none of these types are exposed in the SDK's API, so JSII is happy. This dramatically decreases the public dependency surface area of the SDK and later on we can follow up with installing these dependencies just in time to reduce the package size (related to #2035).

Simplified wingii APIs by removing the concept of `roots` (not used) and the option to not load `deps`. This is not needed now because our SDK is only dependent on `constructs` which we want to load (because we want to allow folks to access its API).

### Example

```js
bring cloud;
bring "constructs" as cx;
bring "aws-cdk-lib" as aws;

resource WingResource {
  init() {
    // access construct node
    log("my id is ${this.node.id}"); 
  } 
}

// reference Construct
let get_path = (c: cx.Construct): str => {
  return c.node.path; 
};

// reference std.Resource
let get_display_name = (r: std.Resource): str? => {
  // reference resource API
  return r.display.title;
};

let q = new aws.aws_sqs.Queue();
let wr = new WingResource();

// wr is-a std.Resource
let another_resource: std.Resource = wr;

// queue and wr are Constructs
log("path of sqs.queue: ${get_path(q)}");
log("path of wing resource: ${get_path(wr)}");

let title = get_display_name(wr) ?? "no display name"; log("display name of wing resource: ${title}");
```

### Follow ups

- #2097
- #2099
- #2098

### Breaking Changes

BREAKING CHANGE: `core.Resource` is now `std.Resource` (as well as other types that were in that file).

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.