fix: unable to resolve extern depending on cwd (#2093)
Take 2 of https://github.com/winglang/wing/pull/2083

tsconfig.json did not have esModuleInterop, which caused `import open from "open"` to not work as expected. I also could not use `import * as open from "open"` like it was originally because mocking that in vitest didn't work. I manually verified the `wing it` works now.

Original PR description below:

---

Fixes #2081

When running wing from the same directory as the wing file, relative extern specs failed to be found in wingc. This was because the WASI preopens didn't understand the given absolute path when when trying to read the same directory as ".". To be completely honest, I'm not sure why it worked from a parent directory.

I added this test to apps/wing instead of hangar, and while I was there I changed it to use vitest instead of jest since the rest of our repo uses it.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.