fix: updates are always checked on cli startup (#2113)
update-notifier is configured to check for updates once a day. We've been checking it on every startup, which means it makes a network call out to npm. On my machine, this reduced the time for arbitrary `wing` commands by ~300ms.

Simplified the update system. There is an argument to skip it and it no longer has any code to do the installation for you.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.