fix(compiler): added missing binding annotations for inflight inits (#2148)
Before this fix we didn't check for the actual refs in the inflight init and just added annotations to all inner clients without details refs (without information for how they were being used in the inflight init). This had the potential to cause permission errors if an inflight init tried to access a resource.
Added an explicit test for this.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.