chore: ensure hangar tests are generated in ci (#2151)
When the generated test files in hangar were removed from git, they then needed to be generated in CI. This wasn't added, so tests haven't really been fully running.

I considered adding a special nx command for this case, but had issues with passing the shard arg. Plus, it means we'd have to install nx in the E2E job when it's really not necessary.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.