feat!: `resource` is now `class` and `class` is now `inflight class` (WIP)
In order to reduce the number of new concepts in Wing and to improve the mental model behind the language, we decided to get rid of the concept of "resource" and instead simply associate a phase modifier with classes, similarly to what we have for functions.

The default phase is preflight, so basically when a class is declared within a preflight scope, it will be a "preflight class". These classes cannot be instantiated within inflight scopes. To declare a class that can be instantiated within an inflight scope, use `inflight class`.

This change merely changes the language, but not the compiler implementation. Internally we still maintain the two concepts, but that's something we should clean up of course.

TODO:
- [ ] Update docs
- [ ] Update language spec

Resolves #2264

BREAKING CHANGE: use `class` instead of `resource` to declare preflight types and use `inflight class` instead of just `class` to declare inflight-only types.