fix(sdk): incorrect construct fqns in tree.json (#2268)
Fixes #811

The construct info mentioned in the original issue (#811) is created by special runtime values that jsii emits in .js files, which were being overwritten by the SDK's double-compilation process. (Because of that, tree.json showed `constructs.Construct` as a fallback value). By removing the double-compiling, the issue is fixed. (This was made possible by upgrading to jsii 5.x in #2065).

The `tree.json` will only show correct values when the SDK has been compiled down to JavaScript, so they don't show in the SDK unit test snapshots (since the SDK uses vitest to hot-load-in TypeScript code). To that end I've added a unit test in the hangar to snapshot a Wing app's tree.json file to show that the fix works.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.