feat!: goodbye `resource` (#2270)
In order to reduce the number of new concepts in Wing and to improve the mental model behind the language, we decided to get rid of the concept of "resource" and instead simply associate a phase modifier with classes, similarly to what we have for functions.

The default phase is preflight, so basically when a class is declared within a preflight scope, it will be a "preflight class". These classes cannot be instantiated within inflight scopes. To declare a class that can be instantiated within an inflight scope, use `inflight class`.

Before             |After
-------------------|-----------------------
`resource Foo { }` | `class Foo { }`
`class Bar { }`    | `inflight Bar { } `


This change merely changes the language, but not the compiler implementation. Internally we still maintain the two concepts, but that's something we should clean up of course.

Resolves #2264

BREAKING CHANGE: use `class` instead of `resource` to declare preflight types and use `inflight class` instead of just `class` to declare inflight-only types.



*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.