chore(sdk): add retention to queue (#2108)
Closes #1213 

Visibilty timeout is still an approximation in the queue simulator. E.g. if timeout is 30s and the function throws an error after 1s, the message will be pushed back to the queue after 31s overall. I think this is fine at the moment, but could be improved in a follow up issue at some time in the future.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.