fix(compiler): preflight objects cannot reference free variables inflight (#2300)
Re-roll of #2211 with some modifications so the implementation keeps inflight clients in their own separate files. I went through a couple iterations to get to this point, here were the potential problems I found with them:
- as @eladb called out, inlining inflight clients into `preflight.js` as implemented in #2205 seems like it could be problematic if clients need to reference other inflight clients (for example, to access static methods)
- another idea I had was to put all inflight clients all in a single file (`preflight.js` and `inflight.js` - cute, right?) - the problem with this is that global variables could conflict with each other, and there might be scoping issues if/when the user starts defining resources at places besides the top level of a wing file

original description:

---

This pull request adds support for capturing [free variables](https://en.wikipedia.org/wiki/Free_variables_and_bound_variables) in resource classes, a feature that allows Wing to generate more expressive code for inflight clients. It implements a new visitor, modifies the code generation, and adds a new test case and a snapshot to verify the functionality.

The approach I've taken differs from the solution outlined in #1447. The original issue suggests that we transform free variables into resource fields. I saw some risks with this approach since all usages of the global value inside inflight constructors and inflight methods would also need to be transformed (for example, changing `global_bucket` to `this.global_bucket`), and tracking the list of these mappings so that they only apply to inflight code, and preventing name conflicts, could be tricky.

Instead, this PR augments jsification so that free variables are effectively also emitted as free variables in the JavaScript client code. 

I'm not sure this approach is any better, but it seems to work just as well for a number of cases I've tried.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.