chore(sdk)!: remove stateful field (#2352)
Removes the `stateful` field from resources in the SDK. This field doesn't serve any functional purpose right now, and every now and then folks ask about it since it's a public API that appears in IDE auto-completions. Let's add it back when we have a use case for it.

BREAKING CHANGE: `stateful` field has been removed as a built-in field of preflight objects

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.