fix(sdk): updating website permission (#2435)
I was testing cloud.Website() and there was a small issue with permissions.
I am uploading the fix, but you can already see the result at this link.

[static website example](https://d24lp7k5o7ycd7.cloudfront.net/index.html)

Another thing is that in AWS it is not possible to know the URL of the website at compile time since the URL is only created when the deployment is completed (and in this case, when CloudFront is finished activating).


*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.