feat!: goodbye snake case (#2468)
In a heavy heart, we are saying goodbye to Wing's royal `snake_case` naming convention in favor of the evil `camelCase`. We realized that the friction it creates with the JavaScript ecosystem is not worth it. Reducing barriers for adopting Wing and making the interop with JavaScript is higher priority in our design principles.

This change removes the case convention logic from the compiler and updates all wing code in this repo accordingly.

Resolves #1049

BREAKING CHANGE: All Wing APIs now use camel-case instead of snake-case (e.g. `queue.addConsumer()` instead of ~`queue.add_consumer()`~) and no case conversion happens in and out of JSII modules.



*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.