chore(compiler): transform inflight closures into preflight classes (#2411)
Fixes #1448

Adds some code in the compiler that inflight closures in the AST into resources. The main goal of this is to unify how we are analyzing resources and inflight closures so that they are treated the same for the purposes of emitting diagnostic errors and for code generation. It should also get us closer to addressing #390.

Highlights:
- Introduces a new "inflight transformer" phase to the compiler, which runs right after the initial parse tree is created, and before type checking. This is only possible because the transform is purely syntactical - it doesn't require any type information.
  - Side note: in earlier iterations of this PR I had this run after type checking, and it pretty much behaves the same, but it required re-type-checking everything the entire tree
- Removes `capture.rs` compiler phase from the compiler, as it was only being used for analyzing inflight closures in the "old way" (which granted all all permissions to all resources it saw). Now, the capture analysis code we have in jsify.rs (which only works on preflight class AST nodes) is used for everything (though some day we might move this into a separate compiler phase...)
- Removes the `core.Inflight` class from the SDK, since we don't need to code-generate it anymore.

Meta comments: there will probably be some regressions lurking here by virtue of behavior that isn't caught by our test suite. I'm also not sure this is the end-game solution for handling inflight closures in the compiler... but it seems like a step in the right direction.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.