chore: ensure volta preloads node/npm before `npm install` (#2611)
Should close #2590

I think this issue is specific to the npm-install github action. The issue isn't a failure to download node/npm, but rather a failure to lock a file. The implication is that the volta shim is being called in parallel and it doesn't really handle it well here. Forcing serial calls upfront to `node` and `npm` should resolve this.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.