fix(sdk): `vm2` is not good enough
The `vm2` module which used to execute inflight functions within the simulator was not cutting it. The sandbox has limitations such as not being able to handle ESM imports, unable to map certain global types due to the vm2 proxy behavior, and more.

To address this, we now esbuild to bundle the function's code *at runtime* (JIT). This is done during the initialization of the simulator cloud function and then reused across invokes. It will likely increase the simulator app start time by a little but this is not something we optimize for at the moment.

Note: since we use `esbuild` through wasm, we needed the `dom` lib in `tsconfig.test.json`.

Resolves #2084
Resolves #2059
Resolves #2088

### Misc

- Fixed a bug in serialization of null objects (`typeof(null)` in javascript is surprisingly an `object`).