fix(compiler): static inflight bindings (#2581)
Allow capturing and binding to work when only static methods are being used. Before this failed:
```wing
bring cloud;

let b = new cloud.Bucket();

class A {
  static inflight do_static() {
    b.put("bla","bla");
  }
}

test "bla" {
  A.do_static(); // Need to capture and bind stuff in `do_static`
}
```
Now it works.

This required supporting `TypeMember` references in the `FieldReferenceVisitor` when analyzing captures and some refactoring there, the tiny added bonus is better spans in the diagnostics. I also had to make the resources base class in the sdk's `_registerBindObject` static so we can call it from a resource's static context.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.