feat: cloud.Secret docs (#2642)
This PR kicks off the work for our resource docs described in https://github.com/winglang/wing/issues/1438.

Docs are stored in the within the SDK's source code, and is copied to the `docs/` directory during the SDK's build. (We need the copied version to be committed to git because our docsite build process expects all files to be there when it starts building the website).

I've put the docs in a new section in the sidebar named "SDK Resources", but open to feedback if it should be under "Concepts" or "Reference" instead.

<img width="1436" alt="Screenshot 2023-05-24 at 4 21 38 PM" src="https://github.com/winglang/wing/assets/5008987/9f37b5ba-dd0c-4139-985d-038adc86500f">

- [x] PR title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] PR description explains motivation and solution
- [x] Tests added
- [x] Docs updated
- [ ] Needs end-to-end tests (`pr/e2e-full` label)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.