fix(compiler): diagnostics missing error messages (#2665)
Fixes an issue where some type checking diagnostics are generated without spans showing where in your code the error is.

In the process I refactored some of the code for generating error messages. It didn't feel like it was as big of an improvement as I expected so I'm open to reverting if the changes seem counterproductive. There's also this `TypeError` struct I saw which doesn't seem to serve any purpose / it's basically the same as `Diagnostic`, so I was also thinking about removing that, but I wasn't sure so I left it as is.

Misc:
- removed some dead (unused) code

----

- [x] PR title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] PR description explains motivation and solution
- [x] Tests added
- [x] Docs updated
- [ ] Needs end-to-end tests (`pr/e2e-full` label)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.