fix: missing completions in if-statement condition (#2666)
Fixes #2374

### Problem

tree-sitter is unable to recognize the following if-statement:

```
let a = MutMap<str> {};
if a.
```

due to missing at least a `{` to start off a block. This means the entire statement becomes an error, causing the reference inside the condition to get thrown away and unavailable to the compiler/lsp.

### Solution

The idea here is to have something similar to the automatic semicolon insertion, but for blocks instead. This means tree-sitter will automatically insert an empty block `{}` when one is expected but unavailable. Like the auto semicolon, this still results in a parse error but we still create a useful AST.

This not only solves the problem for if-statements, but other similar syntactic structures like for-loops.

- [x] PR title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] PR description explains motivation and solution
- [x] Tests added
- [ ] Docs updated
- [ ] Needs end-to-end tests (`pr/e2e-full` label)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.