fix(compiler): missing environment variable runtime error (#2670)
Fixes #1966
Fixes #2082

Addresses a bug in the compiler that resulted in "missing environment variable" errors appearing in some situations where inflight functions were created inside of class constructors.

Take the following example:

```js
class MyProcessor {
  b: cloud.Bucket;
  q: cloud.Queue;

  init() {
    this.q = new cloud.Queue();
    this.b = new cloud.Bucket();

    this.q.addConsumer(inflight (message: str) => {
      log("processing ${message}..");
      this.b.put("file.txt", message);
    });
  }
}
```

`this.q.addConsumer` creates a `cloud.Function` resource to run the inflight code. That cloud resource was originally receiving the minimal permissions needed to call the "put" method on "this.b". This caused an issue because in the current way the compiler generates inflight code, it is bundling and instantiating the entire client of "this" - i.e. the inflight client of MyProcessor.

This causes issues because the inflight client code of MyProcessor is a JavaScript class that looks like this:

```js
  class MyProcessor {
    constructor({ b, q }) {
      this.b = b;
      this.q = q;
    }
  }
```

... and declaring `q` requires instantiating a `cloud.Queue` client, which requires these extra environment variables.

The current fix I've added is to grant "inflight init" permissions to the entirety of `this` despite the fact that only the child (`this.b`) is used.

In the future we might want to generate inflight code differently so only `this.b` has to be bundled, but I think it might require more work.

- [x] PR title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] PR description explains motivation and solution
- [x] Tests added
- [x] Docs updated
- [ ] Needs end-to-end tests (`pr/e2e-full` label)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.