fix: build process cwd is wrong in the console (#2687)
The Wing Console fails when using the `@winglang/sdk` package (specifically, the Simulator) because it uses the `esbuild-wasm` package in ways that bundlers can't manage properly.

This change sets the cwd of the `esbuild-wasm` spawned process to `__dirname`, and removes the `require(require.resolve("esbuild-wasm"))` to a normal `require("esbuild-wasm")`. Since the process' cwd is now the same as `libs/wingsdk/src/shared/bundling.ts` directory, Node will naturally find `esbuild-wasm`, and our Console's dist files, which include the SDK, will work as expected as well. This is necessary because the cwd in the Console environment is the root directory (`/`).

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.