fix(compiler): calls and `if let` expressions evaluated twice
When jsifying call and `if let` expressions we generate code that uses the original expression multiple times. If this expression is e.g. a call or a `new` statement, there could be side effects to the fact that it gets called.

To fix the function call case, I've changed the strategy taken in #2580. Instead of testing whether an object is a function or an object with a `handle` method, we are now returning a "function proxy" from the constructor of the generated handle class inflight. This technique allows the returned object to behave both as a function and as an object:

```js
class MyHandler {
  constructor() {
    const $obj = (...args) => this.handle(...args);
    Object.setPrototypeOf($obj, this);
    return this;
  }

  async handle() { ... }
}
```

This cleans up *a lot* of boilerplate from the call sites because now they don't care if an object a normal function or a handle function. Neat!

The `if let` fix was easier: basically just use the `$IF_LET_VALUE` variable inside the condition instead of replicating the expression.