fix!: unable to use `test` without `bring cloud` (#2674)
Move `cloud.Test` to `std.Test` based on the approach that `std` includes types that are implicitly imported ("brought?"). But... for some reason we have special-cases for *every* type in `std` when we jsifiy, so now we are implicitly converting references to `std` during jsification to `$stdlib.std`.

As a follow up we might be able to get rid of many of the special cases we have for `std` types, but I'll save that for a future refactor.

## Fixes

Fixes #2521
Fixes #2619
Fixes #2598
Fixes #2702

## Misc

- Improve diagnostics for internal compiler errors (throw an error instead of `process.exit()`!).
- Only use diagnostic error formatter if `span.file_id` is not empty. The reason we didn't have diagnostics in the case where `cloud` was not brought was because it's span was empty.


## Checklist

- [x] PR title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] PR description explains motivation and solution
- [x] Tests added
- [x] Docs updated
- [x] Needs end-to-end tests (`pr/e2e-full` label)

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.