fix(sdk): aws test when invoke a function that returns null (#2688)
The PR #2587 modification fixed tests that were returning strings when executed on AWS, but it created a problem when the function didn't return a value (null). This PR validates this return and returns an empty string in these cases.

The test `examples/tests/sdk_tests/function/invoke.w` validates this modification when integrated with AWS.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.