fix(vscode): language server cannot recover from panics (#2738)
Normally when the language server panics, the whole thing crashes and you need to restart your editor. This adds a built-in mechanism to avoid exiting completely. It instead lets you know and waits until the wing file is changed before trying again.

The method is pretty naive and effectively resets the entire wasm memory. I did this because the thread local storage has a poisoned refcell and I couldn't figure out how to clear it. I think a full reset is avoidable but this current solution is ok.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.