fix: unable to call static namespaced methods from inflight
Static namespaced methods, such as the ones vended as part of the Wing SDK have the form `namespace.Class.method()`. In order to be able to call this method we need to capture the class object at runtime (also known as the "inflight type").

Since this is a dot delimited string (`ns.Class`), we need to mangle it in order to be able to pass it through the capture mechanism. So we represent these types as `namespace_Class` within the inflight code and call then call `namespace.Class._toInflightType()` on the preflight side (which is how static methods are supported for non-namespaced types).

Then, we must add a `_toInflightType()` static method to `std` and `util` classes, which will `require()` the correct SDK javascript file and return the type itself. Added a test to make sure this is covered for future SDK classes.

### Resolves

* Fixes #2576

### Misc

* Added `WING_TARGET` to cloud function environments (for all targets).
* Created an initial set of SDK tests for `std` types.
* Some renames in `jsify.rs` to normalize the mental model a little.
* Switch the implementation of `Boolean.fromJson` from a macro to an actual method (just for the fun of it).