fix: unable to call static namespaced methods from inflight (#2783)
Static namespaced methods, such as the ones vended as part of the Wing SDK have the form `namespace.Class.method()`. In order to be able to call this method we need to capture the class object at runtime (also known as the "inflight type").

Since this is a dot delimited string (`"namespace.Class"`), we need to mangle it in order to be able to pass it through the capture mechanism. So we represent these types as `"namespace_Class"` within the inflight code and then call `namespace.Class._toInflightType()` on the preflight side (which is how static methods are supported for non-namespaced types).

Then, we must add a `_toInflightType()` static method to `std` and `util` classes, which will `require()` the correct SDK javascript file and return the type itself. Added a test to make sure this is covered for future SDK classes.

### Resolves

* Fixes #2576

### Follow ups

* https://github.com/winglang/wing/issues/2782
* https://github.com/winglang/wing/issues/2785

### Misc

* Added `WING_TARGET` to cloud function environments (for all targets).
* Created an initial set of SDK tests for `std` types.
* Some renames in `jsify.rs` to normalize the mental model a little.
* Switch the implementation of `Boolean.fromJson` from a macro to an actual method (just for the fun of it).
* Reduced progress logs from `wing test` by default and add a `--progress` option to enable.
* Prelude: https://github.com/winglang/wing/pull/2799

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.