feat(sdk): adding awscdk test command (#2612)
This pull request adds the `awscdk` target to the test command. (Everyone :heart:'s a long-running test, right?)

You can run a cdk test with this steps:
* set the CDK_STACK_NAME environment variable
* run tests with awscdk target

```bash
$ export CDK_STACK_NAME=test-my-cdk-stack
$ wing test -t awscdk app.w
```

Two things should be noted. In the ENV_WING_TEST_RUNNER_FUNCTION_ARNS variable, I had to remove the underscore because the CloudFormation output doesn't accept special characters. And to capture the ARN of the test function, I had to create an output.json file in the directory of the created object. But don't worry, I delete the file in the end when the stack is being removed.



*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.