TODO:  - [ ] There's some failing scenarios still
Add initial support for hover documentation for classes, functions and symbols.

## Implementation Notes

Propagate the JSII docs throughout the type system. This is currently done for function signatures and parameters as well as classes.

Created a `Documented` trait and implement it for various types in order to allow polymorphic doc generation.
Removed all the spans from the synthetic syntax produced by the `ClosureTransformer` since they shadowed the actual symbols in the source file and that caused the hover visitor to ignore any statements within the transformed closure.

Rewrite HoverVisitor to find the relevant element and use the new trait to render docs.

Added `name` to `FunctionParameter` so it will be possible to display the parameter names in docs.

Added the desugaring phase to the LSP so that closure will be transformed appropriately.

## Misc

* Removed special handling of the `core.Inflight` type which no longer exists.
* Add `full_path_str` to `UserDefinedType` to make it easier to produce a nested string for lookups.
* Extended visitor to also allow subscribing to symbol visits (low level).
* Removed the name of the interface when showing closure signatures. This fixes #2786.
* Do not apply spans to default initializers of classes.