fix(compiler): panic when trying to reference an unknown field
Fixes #2602 by avoiding the creation of namespaces blindly, even if a type doesn't exist in that namespace.

Additionally fixes #2877 which was caused by mistakenly associating its symbol environment with the parent submodule, but that's not actually what we want. If I couldn't find `String` in `std.foo`, I don't want to look up string in `std`.