fix(compiler): type check assertion is ignored (#2868)
I realized I introduced a small bug in my last PR [here](https://github.com/winglang/wing/pull/2860) - inside the main compiler pipeline (inside `compile.rs`) I added a call to `check_all_exprs_type_checked`, but didn't actually check the result. When we do check the result, it claims many AST nodes that are not type checked because it's looking at all `Expr`'s that have ever been created - but the reason for that is that some `Expr`'s nodes get thrown away. For example, in this snippet of Wing code:

```js
enum Foo { A, B }

let a = Foo.A;
```

We initially parse the expression in the last statement as something like

```js
Expr(Ref::InstanceMember(Expr(Ref::Identifier("Foo")), "A"))
```

But it's converted into this during type checking

```js
Expr(Ref::TypeMember(UserDefinedType("Foo"), "A"))
```

So, to fix this I just went back to the old strategy of traversing the final tree.

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.