fix(compiler): panic when trying to reference an unknown field (#2878)
Fixes #2602 by avoiding the creation of namespaces blindly, even if a type doesn't exist in that namespace.

Additionally fixes #2877 which was caused by mistakenly associating its symbol environment with the parent submodule, but that's not actually what we want. If I couldn't find `String` in `std.foo`, I don't want to look up string in `std`. The diagnostics emitted in this case is still weird, so I created #2876 to address this.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [x] Tests added (always)
- [x] Docs updated (only required for features)
- [x] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.