feat(vscode): show docs on hover (#2534)
Add initial support for hover documentation for classes, functions and symbols.

<img width="649" alt="Screenshot 2023-06-09 at 16 24 17" src="https://github.com/winglang/wing/assets/598796/140b5c79-10c3-4ea3-8fce-3e4935dbb9b7">

<img width="368" alt="Screenshot 2023-06-09 at 16 24 35" src="https://github.com/winglang/wing/assets/598796/f6b66b4c-c64d-40c0-9bc6-2ce992a79bb3">

<img width="438" alt="Screenshot 2023-06-09 at 16 25 34" src="https://github.com/winglang/wing/assets/598796/b8d6d75d-6a2b-4355-b1db-9bfad583e4a0">

Propagate the JSII docs throughout the type system. This is currently done for function signatures and parameters as well as classes.

Created a `Documented` trait and implement it for various types in order to allow polymorphic doc generation.
Removed all the spans from the synthetic syntax produced by the `ClosureTransformer` since they shadowed the actual symbols in the source file and that caused the hover visitor to ignore any statements within the transformed closure.

Rewrite HoverVisitor to find the relevant element and use the new trait to render docs.

Added `name` to `FunctionParameter` so it will be possible to display the parameter names in docs.

Added the desugaring phase to the LSP so that closure will be transformed appropriately.

* Removed special handling of the `core.Inflight` type which no longer exists.
* Add `full_path_str` to `UserDefinedType` to make it easier to produce a nested string for lookups.
* Extended visitor to also allow subscribing to symbol visits (low level).
* Removed the name of the interface when showing closure signatures. This fixes https://github.com/winglang/wing/issues/2786.
* Do not apply spans to default initializers of classes.
* Fix #1569 by adding parameter names to function signatures.

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.