chore: remove volta from ci (#2891)
Fixes #2590

Volta was causing too many intermittent issues in CI. I kept it in the docs locally because I still think it's really great, but I'm open to removing it from there as well if it's causing enough issues for others. 

Removed from the devcontainer as well, which helps because it let's us use `postCreateCommand` and `updateContentCommand` which will be great if we want to start doing prebuilds. Verified the change locally.

The only real downside of this change is that now the node/npm version is duplicated, but I think the tradeoff is worth it.

## Checklist

- [x] Title matches [Winglang's style guide](https://docs.winglang.io/contributors/pull_requests#how-are-pull-request-titles-formatted)
- [x] Description explains motivation and solution
- [ ] Tests added (always)
- [ ] Docs updated (only required for features)
- [ ] Added `pr/e2e-full` label if this feature requires end-to-end testing

*By submitting this pull request, I confirm that my contribution is made under the terms of the [Monada Contribution License](https://docs.winglang.io/terms-and-policies/contribution-license.html)*.