feat: allow referencing properties of CDK constructs
It is now possible to reference properties of Terraform CDK constructs from inflight context. For example:

```js
bring "@cdktf/provider-aws" as aws;
let b = new aws.s3Bucket.S3Bucket();

inflight () => {
  log(b.arn);
};
```

In order to support this, we need the lifting and binding mechanism to have information about which properties are being accessed. To that end, I've extended this API to accept an additional `props` array with the names of properties being bound/lifted.

I changed how `FieldReferenceVisitor` so that if it finds a property reference (e.g. `this.bucket.arn` or `bucket.arn`), it will treat `arn` as the `op`, and we merge this information with the list of captured variables/fields.

Tweaked the `registerBind()` method to allow any `op` that matches a preflight property.

* Fixed a bug in how we lift tokens. CDKTF produces a new token every time a tokenized property is retrieved (e.g. `bucket.arn`). And since we used this token string every time. Since we use the token ID to produce an environment variable, this created a situation where `lift()` and `bind()` had a different environment variable name.
* Update the `Debug Wing Compiler` launch configuration to use the currently open file as the source for debugging.
* Remove the `self_client_path` from the generated `_toInflightType()` methods and just embed the path directly into the inline code.